De-linting, and more moving
Shifting some more stuff around to make things easier to work on.  This
also removes some bad syntax that snuck in due to the last refactor.

This pass actually got run in a browser, so it seems to work!  Added a
pre-constructed global `reckoning` instance as well for fast access to
the helpers (map etc).  I'd prefer for this to work off the main
constructor so there aren't two namespaces, but that can come later.

Next up is the fun stuff - formatting and generating calendars. :)