Working Date Ranges
Renamed the internal 'Range' constructor to 'DateRange' to prevent
naming conflicts/confusion with the upcoming native Range.  Refactored
the 'mapRange' function on the main Reckoning object to call the
DateRange constructor, and fleshed-out the range API.  It should now be
fully-functional!

I've been saying it for a while, but next steps are to start on
visualizations. :)