Range Simplifications
Using dynamic ranges should make it easier to build responsive
calendars, so mapping fixed dates between a `fromDate` and `toDate` has
been relegated to an optional utility (currently enabled with an
'between' parameter flag that would be used on the `mapRange` Reckoning
function.  (This could be potentially revisited to make the
functionality more clear - should be addressed prior to v0.0.1).

This commit also removes the closure from the main Reckoning constructor
- it seems unnecessary now that it's correctly broken-up and no longer
  the monolithic beast that it was when prototyping.

Additionally, the 'calendar' api underwent some minor changes (which
doesn't matter much yet because we're not yet building a calendar...).