Calendar Views
This adds the long-awaited mithril templating that will actually display
a calendar.  Yay!  Now that we can actually see them, it's kind of
obvious that there was a fundamental flaw in the original 'mapDays'
function, so for now we've just added a small offset.

It ain't yet pretty, but we're going somewhere here.  Strap in. :)