Calendar View Examples
This cleans up the examples with some quick-n-dirty styling to get the
ball rolling.

Next steps should be a pretty basic theme for the calendar (because hey
- this is a visualization library, right?)

Putting together the demo is starting to reveal that we may want to
enable some default ranges if nothing is specified, but I'm still
leaning towards giving developers free reign here and thus leaving
Reckoning with a broader use-case.