More Robust Format Fallback
Safari doesn't support locale options, so that kind of stinks.  Making
localization work by default is a key goal for this project, but some of
that may be outside the scope of where we can get for 0.0.1.  In the
meantime, we can make sure that calendar views look consistent in
english across browsers that don't currently support native localized
strings.  This rounds-out the `format` fallback, and improves the month
and day mapping so that they'll shorten strings according to the
constructor settings (or passed-in options).