Improve Accessibility
This adds a keypress endpoint for each day cell, along with a number of
other best-practice html attributes to help each calendar 'look' better
to assistive devices.  Days in every month are also tracked by date to
make it easer to map focus rules etc for different use cases.  This
required a bit of refactoring but the functionality that comes from this
should be worth it. :)