Improved Demo Page
This makes some adjustments to the primary reckoning styles that became
apparant after using it, fixes a hidden problem with falsy checks for 0
values that future tests would catch anyway and... updates the demo page
to be way more useful and visually attractive.  I think it shows off the
power of Reckoning a lot better now.