Calendar Controls and Model Refactor
Support for displaying calendar controls which makes all of the crazy
data logic worthwhile, and some adjustments to clarify the difference
between data (model), and functional elements (like parse, format, etc).

It got a bit confusing with everything lumped together, especially since
for the most part the 'controllers' are the top-level constructors.

Also, trying to use local vars more instead of long direct references to
other related controllers (which may in fact be a bad practice, but it's
just so darn easy this way).

Hopefully things are easier to grasp going forward.