Fix Redraw Performance
I made a bit of a silly mistake by calling an expensive date formatting
function directly from the view rather than caching the result in a
viewmodel.  This MASSIVELY improves the redraw time when interacting
with the calendar.

Oh, and I fixed the rogue dashes in the gruntfil that prevented watch
from working.  derp.