Add View Wrapper for Months
Experimenting with Reckoning in the 'real-world' quickly revealed that
multi-month calendars need a bit more flexibility.  This adds a view
wrapper for months that will pass back the calendar instance along with
the original view.  It's up to the developer to actually use that view,
or throw it away like GARBAGE. :)

This also modifies the `dayView` to return the full controller.  The
difference between `monthView` and `dayView` is that the former is a
wrapper and the latter is an inner view.  This could potentially be
confusing, but it seems reasonable this way since a totally custom
`monthView` is unlikely to use a the built-in `dayView` method so the
only thing that needs to be consistent is that the entire controller is
returned to each view type.