Fix Calendar VM
It was nagging me that the Calendar constructor was modifying the parent
view.  This removes that hacky move and replaces it with a slightly more
palateable method.  Some of this it starting to feel a bit clunky,
because every time I think about updating the 'starting day' I would
need to re-map the weekdays and then update the month list and ...
hrm...