Custom 'proppy' to make Reckoning dependency-free.
Using a custom gettersetter has a side-benefit of allowing for internal
callbacks to update related data.  This is often an undesirable pattern,
but when it comes it calendars and handling the varaity of options that
can modify different data points it's kind of unavoidable.  The callback
prop pattern could now be extended to locales and string options, which
is fun. :)

Also added manual redraw calls (which default to noop), as some
document-oriented frameworks will likely need that in order to update
correctly.