Removing global 'rk' shortcut, `toJSON` support.
Reckoning should always be used by the core namespace in order to
prevent confusion.  These changes make it easier to actually do that.
Some of the core functions were moved out to the main `Reckoning`
namespace, the `locale` function was made a bit more robust, and the
core function returns something interesting when called directly (which
may change - I'm not yet sure what I want to do with that yet.)

`toJSON` is a standard method that should make it easier to work with
Reckoning isomorphically (i.e. in node and in the browser).

This moves Reckoning closer to feeling like a 'standard'.