Basic React Implementation
This creates a 'Calendar' react class which is publically exposed as
`Reckoning.react.Calendar`.  There's a lot of click-handler duplication
that Reckoning typically handles internally but needs to be passed up
and down the react tree in order to porpopgate correctly.  I'd like to
know if this could be cleaned-up, but this seems like the most stable
way to handle things - basically the only state that a Reckoning
calendar should actually care about is the current month/year
positioning - everything else should be handled by a parent state. (e.g.
ranges).