Housecleaning, Tests, v0.1.0
This removes the 'Timeline' stubs, adds a bunch of sorely-needed tests,
and performs some general cleanup in preperation for the release of
v0.1.0.  Reckoning is now in a generally stable state and forthcoming
changes shouldn't necesetate any major changes to the overall API.  The
important core stuff is now covered by proper functional tests, and I
finally figured out a way to remove that stupid `calVM` from the views.