🎓  Pre-bind Event Handlers
Binding each event handler in different ways in the view is bad, and
having to re-bind on every render is worse.  Reckoning is intended to be
a calendar state generator, but it's hard to draw the line and leave
things to the view that really shouldn't be the view's concern.  This
performs a slightly dirty optimization where each date's handlers are
stored to a map on the calendar - that way when dates are re-rendered,
the bindings don't need to be regenerated every time.